+++ docs/qa/scenarios/LP-cancel-vs-kill.md
@@ -11,9 +11,11 @@ bug_ids:
 fix_status:
 retest_status: pass
 fix_commits:
-evidence: looprun-db9; looprun-1e5; looprun-08b; looprun-d7f; /Users/pedronauck/dev/qa-labs/compozy-loop-node-lifecycle-20260803-191237-281307-lab/qa-artifacts/qa;/Users/pedronauck/dev/qa-labs/compozy-loop-agent-ownership-r2-20260806-040706-936266-lab/qa-artifacts/qa/evidence/loop-cancel-draining.json;/Users/pedronauck/dev/qa-labs/compozy-loop-agent-ownership-r2-20260806-040706-936266-lab/qa-artifacts/qa/evidence/loop-cancel-latest.json
-last_report: docs/qa/reports/2026-08-06-loop-agent-ownership.md
+evidence: looprun-db9; looprun-1e5; looprun-08b; looprun-d7f; /Users/pedronauck/dev/qa-labs/compozy-loop-node-lifecycle-20260803-191237-281307-lab/qa-artifacts/qa;/Users/pedronauck/dev/qa-labs/compozy-loop-agent-ownership-r2-20260806-040706-936266-lab/qa-artifacts/qa/evidence/loop-cancel-draining.json;/Users/pedronauck/dev/qa-labs/compozy-loop-agent-ownership-r2-20260806-040706-936266-lab/qa-artifacts/qa/evidence/loop-cancel-latest.json;docs/qa/reports/2026-08-31-loop-result-fix.md
+last_report: docs/qa/reports/2026-08-31-loop-result-fix.md
 overlaps: LP-016
 ---
 
 acceptance-walk: Cancel one active run and kill another through Web, then repeat and race the run and node verbs through structured CLI and HTTP. Confirm the cooperative drain versus immediate session stop, exact terminal causes, one terminal effect, deterministic loser responses, fresh-read parity, and absence of every retired stop control or route.
+
+QA impact 2026-08-31: targeted public-daemon E2E replay passed after fixing the cancellation race where a run-owned Goal binding could be created after the cancellation request timestamp. The binding failure and cleanup obligation now share a terminal timestamp that never precedes binding creation; the wider cancel-versus-kill contract remains covered by the prior acceptance walk.
