+++ web/package.json
@@ -79,7 +79,6 @@
         "@types/react-dom": "^19.2.5",
         "@vitejs/plugin-react": "^6.1.0",
         "babel-plugin-react-compiler": "^1.0.0",
-        "emojibase-data": "17.0.0",
         "fake-indexeddb": "^6.2.5",
         "jsdom": "^29.1.1",
         "lucide-static": "1.34.0",
