+++ web/src/systems/tasks/components/task-overview-panel.tsx
@@ -4,12 +4,10 @@ import type { ComponentPropsWithoutRef } from "react";
 import { Button, cn, DescriptionCard, Section } from "@compozy/ui";
 
 import type { TaskDetailView, TaskRun, TaskTimelineItem } from "../types";
-import { latestTaskRun } from "../lib/task-run-presentation";
-import { useTaskRunResult } from "../hooks/use-task-run-result";
 import { TaskActivityItem } from "./task-activity-item";
 import { TaskDependenciesSection } from "./task-dependencies-section";
 import { TaskNowStrip, type TaskNowStripHandlers } from "./task-now-strip";
-import { TaskResultSection } from "./task-result-section";
+import { TaskResultSection, type TaskResultPageController } from "./task-result-section";
 import { TaskSubtasksSection } from "./task-subtasks-section";
 import { TASK_RESULT_ANCHOR_ID } from "./task-overview-constants";
 
@@ -24,31 +22,19 @@ export interface TaskOverviewPanelProps extends ComponentPropsWithoutRef<"div"> 
   nowPending?: Parameters<typeof TaskNowStrip>[0]["pending"];
   onViewAllActivity: () => void;
   activeRunElapsed?: string;
+  completedResult?: {
+    external?: TaskResultPageController;
+    run: TaskRun;
+  };
 }
 
-function CompletedTaskResult({
-  runs,
-  workspaceId,
+function CompletedRunResult({
+  external,
+  run,
 }: {
-  runs: readonly TaskRun[];
-  workspaceId: string;
+  external?: TaskResultPageController;
+  run: TaskRun;
 }) {
-  const lastCompleted = latestTaskRun(runs, "completed");
-  if (!lastCompleted) return null;
-
-  return (
-    <CompletedRunResult key={lastCompleted.id} run={lastCompleted} workspaceId={workspaceId} />
-  );
-}
-
-function CompletedRunResult({ run, workspaceId }: { run: TaskRun; workspaceId: string }) {
-  const external = useTaskRunResult({
-    resultBytes: run.result_bytes ?? 0,
-    resultRef: run.result_ref ?? "",
-    runId: run.id,
-    workspaceId,
-  });
-
   return (
     <TaskResultSection
       data-testid="tasks-detail-result-section"
@@ -79,6 +65,7 @@ export function TaskOverviewPanel({
   nowPending,
   onViewAllActivity,
   activeRunElapsed,
+  completedResult,
   className,
   ...props
 }: TaskOverviewPanelProps) {
@@ -102,8 +89,12 @@ export function TaskOverviewPanel({
         runs={runs}
       />
 
-      {record.status === "completed" ? (
-        <CompletedTaskResult runs={runs} workspaceId={record.workspace_id ?? ""} />
+      {record.status === "completed" && completedResult ? (
+        <CompletedRunResult
+          external={completedResult.external}
+          key={completedResult.run.id}
+          run={completedResult.run}
+        />
       ) : null}
 
       <Section data-testid="tasks-detail-description" label="Description">
